+++ dash-renderer/src/utils/callbacks.ts
@@ -1,0 +1,15 @@
+import { ICallbacksState } from '../reducers/callbacks';
+
+export const getPendingCallbacks = ({
+    executed,
+    executing,
+    prioritized,
+    requested,
+    watched
+}: ICallbacksState) => [
+    ...requested,
+    ...prioritized,
+    ...executing,
+    ...watched,
+    ...executed
+];
